Refactoring - Moved UserAuthDao methods from UserDao to UserAuthDao class
