+++ quora-service/src/main/java/com/upgrad/quora/service/business/QuestionBusinessService.java
@@ -1,6 +1,7 @@
 package com.upgrad.quora.service.business;
 
 import com.upgrad.quora.service.dao.QuestionDao;
+import com.upgrad.quora.service.dao.UserAuthDao;
 import com.upgrad.quora.service.dao.UserDao;
 import com.upgrad.quora.service.entity.QuestionEntity;
 import com.upgrad.quora.service.entity.UserAuthEntity;
@@ -22,6 +23,9 @@ public class QuestionBusinessService {
     private UserDao userDao;
 
     @Autowired
+    private UserAuthDao userAuthDao;
+
+    @Autowired
     private QuestionDao  questionDao;
 
 
@@ -36,7 +40,7 @@ public class QuestionBusinessService {
     public QuestionEntity createQuestion(QuestionEntity newQuestion, final String authorizationToken)
             throws AuthorizationFailedException{
 
-       UserAuthEntity userAuthEntity = userDao.getUserAuth(authorizationToken);
+       UserAuthEntity userAuthEntity = userAuthDao.getUserAuth(authorizationToken);
         if(userAuthEntity == null){
             throw new AuthorizationFailedException("ATHR-001","User has not signed in");
         } else{
@@ -61,7 +65,7 @@ public class QuestionBusinessService {
      * @throws AuthorizationFailedException
      */
     public List<QuestionEntity> getAllQuestions(final String authorizationToken) throws AuthorizationFailedException{
-        UserAuthEntity userAuthEntity = userDao.getUserAuth(authorizationToken);
+        UserAuthEntity userAuthEntity = userAuthDao.getUserAuth(authorizationToken);
         if(userAuthEntity == null){
             throw new AuthorizationFailedException("ATHR-001","User has not signed in");
         } else{
@@ -88,7 +92,7 @@ public class QuestionBusinessService {
             QuestionEntity updatedQuestion, final String authorizationToken)
             throws AuthorizationFailedException, InvalidQuestionException{
 
-        UserAuthEntity userAuthEntity = userDao.getUserAuth(authorizationToken);
+        UserAuthEntity userAuthEntity = userAuthDao.getUserAuth(authorizationToken);
         if(userAuthEntity == null){
             throw new AuthorizationFailedException("ATHR-001","User has not signed in");
         } else{
@@ -124,7 +128,7 @@ public class QuestionBusinessService {
             final String questionUuid, final String authorizationToken)
             throws AuthorizationFailedException, InvalidQuestionException{
 
-        UserAuthEntity userAuthEntity = userDao.getUserAuth(authorizationToken);
+        UserAuthEntity userAuthEntity = userAuthDao.getUserAuth(authorizationToken);
         if(userAuthEntity == null){
             throw new AuthorizationFailedException("ATHR-001","User has not signed in");
         } else{
@@ -158,7 +162,7 @@ public class QuestionBusinessService {
      */
     public List<QuestionEntity> getAllQuestionsByUser(final String uuidOfUser, final String authorizationToken)
         throws AuthorizationFailedException, UserNotFoundException{
-        UserAuthEntity userAuthEntity = userDao.getUserAuth(authorizationToken);
+        UserAuthEntity userAuthEntity = userAuthDao.getUserAuth(authorizationToken);
         if(userAuthEntity == null){
             throw new AuthorizationFailedException("ATHR-001","User has not signed in");
         } else{
