+++ quora-service/src/main/java/com/upgrad/quora/service/dao/UserDao.java
@@ -1,6 +1,5 @@
 package com.upgrad.quora.service.dao;
 
-import com.upgrad.quora.service.entity.UserAuthEntity;
 import com.upgrad.quora.service.entity.UserEntity;
 import org.springframework.stereotype.Repository;
 
@@ -81,37 +80,6 @@ public class UserDao {
         }
     }
 
-    /**
-     * Creates a new entry in the user_auth table
-     * @param userAuthEntity
-     * @return
-     */
-    public UserAuthEntity createUserAuth(final UserAuthEntity userAuthEntity){
-        entityManager.persist(userAuthEntity);
-        return userAuthEntity;
-    }
-
-    /**
-     * Fetches the record from UserAuthEntity using access token
-     * @param accessToken
-     * @return
-     */
-    public UserAuthEntity getUserAuth(final String accessToken){
-        try {
-            return entityManager.createNamedQuery("userAuthByAccessToken", UserAuthEntity.class)
-                    .setParameter("accessToken", accessToken).getSingleResult();
-        } catch (NoResultException nre) {
-            return null;
-        }
-    }
-
-    /**
-     * Updates an existing UserAuthEntity record.
-     * @param userAuthEntity
-     */
-    public void updateUserAuth(final UserAuthEntity userAuthEntity){
-        entityManager.merge(userAuthEntity);
-    }
 
     /**
      * Delete the userEntity
